Moving phone home stuff out of scan config
